refactor: SessionRepositoryImpl 제거
