+++ tomcat/src/main/java/org/apache/catalina/Manager.java
@@ -4,14 +4,12 @@ import jakarta.servlet.http.HttpSession;
 import java.io.IOException;
 
 /**
- * A <b>Manager</b> manages the pool of Sessions that are associated with a
- * particular Container.  Different Manager implementations may support
- * value-added features such as the persistent storage of session data,
- * as well as migrating sessions for distributable web applications.
+ * A <b>Manager</b> manages the pool of Sessions that are associated with a particular Container.  Different Manager
+ * implementations may support value-added features such as the persistent storage of session data, as well as migrating
+ * sessions for distributable web applications.
  * <p>
- * In order for a <code>Manager</code> implementation to successfully operate
- * with a <code>Context</code> implementation that implements reloading, it
- * must obey the following constraints:
+ * In order for a <code>Manager</code> implementation to successfully operate with a <code>Context</code> implementation
+ * that implements reloading, it must obey the following constraints:
  * <ul>
  * <li>Must implement <code>Lifecycle</code> so that the Context can indicate
  *     that a restart is required.
@@ -31,18 +29,13 @@ public interface Manager {
     void add(HttpSession session);
 
     /**
-     * Return the active Session, associated with this Manager, with the
-     * specified session id (if any); otherwise return <code>null</code>.
+     * Return the active Session, associated with this Manager, with the specified session id (if any); otherwise return
+     * <code>null</code>.
      *
      * @param id The session id for the session to be returned
-     *
-     * @exception IllegalStateException if a new session cannot be
-     *  instantiated for any reason
-     * @exception IOException if an input/output error occurs while
-     *  processing this request
-     *
-     * @return the request session or {@code null} if a session with the
-     *         requested ID could not be found
+     * @return the request session or {@code null} if a session with the requested ID could not be found
+     * @throws IllegalStateException if a new session cannot be instantiated for any reason
+     * @throws IOException           if an input/output error occurs while processing this request
      */
     HttpSession findSession(String id) throws IOException;
 
@@ -52,4 +45,5 @@ public interface Manager {
      * @param session Session to be removed
      */
     void remove(HttpSession session);
+
 }
