+++ tomcat/src/main/java/org/apache/catalina/connector/Connector.java
@@ -91,4 +91,5 @@ public class Connector implements Runnable {
     private int checkAcceptCount(final int acceptCount) {
         return Math.max(acceptCount, DEFAULT_ACCEPT_COUNT);
     }
+
 }
