+++ tomcat/src/main/java/org/apache/catalina/startup/Tomcat.java
@@ -23,4 +23,5 @@ public class Tomcat {
             connector.stop();
         }
     }
+
 }
