--- tomcat/src/main/java/org/apache/coyote/http11/request/HttpRequest.java
@@ -1,19 +1,0 @@
-package org.apache.coyote.http11.request;
-
-import org.apache.coyote.http11.request.body.RequestBody;
-import org.apache.coyote.http11.request.headers.RequestHeader;
-import org.apache.coyote.http11.request.line.RequestLine;
-
-public class HttpRequest {
-
-    private final RequestLine requestLine;
-    private final RequestHeader requestHeader;
-    private final RequestBody requestBody;
-
-    public HttpRequest(RequestLine requestLine, RequestHeader requestHeader, RequestBody requestBody) {
-        this.requestLine = requestLine;
-        this.requestHeader = requestHeader;
-        this.requestBody = requestBody;
-    }
-
-}
