+++ tomcat/src/main/java/org/apache/coyote/http11/request/headers/RequestHeader.java
@@ -10,6 +10,7 @@ import static java.util.Arrays.asList;
 import static java.util.Collections.EMPTY_LIST;
 
 public class RequestHeader {
+
     private static final String COOKIE_HEADER = "Cookie";
 
     private final Map<String, List<String>> headers;
