+++ tomcat/src/main/java/org/apache/coyote/http11/response/ResponseEntity.java
@@ -34,4 +34,5 @@ public class ResponseEntity {
     public Cookie getHttpCookie() {
         return cookie;
     }
+
 }
